+++ docs/src/nav-links.js
@@ -239,6 +239,11 @@ export default [
         href: `${BASE_URL}#/docs/utilities/cda-url-provider`,
       },
       {
+        id: "get-precision",
+        text: "Get Precision",
+        href: "/docs/utilities/get-precision",
+      },
+      {
         id: "use-debounce",
         text: "Debounce",
         href: `${BASE_URL}#/docs/utilities/use-debounce`,